refactor: 패키지 구조 수정 - HTTP 요청을 받는 부분을 모두 WebServer 패키지로 통합합니다.
