+++ src/main/java/codesquad/webserver/http/type/ContentType.java
@@ -1,4 +1,4 @@
-package codesquad.http.type;
+package codesquad.webserver.http.type;
 
 import java.util.Arrays;
