+++ src/main/java/codesquad/webserver/http/type/HttpStatus.java
@@ -1,4 +1,4 @@
-package codesquad.http.type;
+package codesquad.webserver.http.type;
 
 public enum HttpStatus {
     // 1xx Informational
