+++ src/main/java/codesquad/webserver/http/type/StartLine.java
@@ -1,4 +1,4 @@
-package codesquad.http.type;
+package codesquad.webserver.http.type;
 
 import java.util.Map;
