+++ src/test/java/codesquad/webserver/http/parser/HeaderParserTest.java
@@ -1,8 +1,9 @@
-package codesquad.http.parser;
+package codesquad.webserver.http.parser;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import codesquad.http.type.HttpHeader;
+import codesquad.webserver.http.parser.HeaderParser;
+import codesquad.webserver.http.type.HttpHeader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
